Add validation to proc-macro-derive
We need to check this on proc-macro-derive because cfg may reduce the
fields. On the other hand, if we check this only on proc-macro-derive,
it may generate unhelpful error messages. So, we need to check this on
both proc-macro-attribute and proc-macro-derive.
